Use md5 hash values for file content in FSCache

+++ src/store/cache/FSCache.tsx
@@ -1,21 +1,22 @@
-import React, { createContext, ReactNode, useEffect } from 'react';
-import { PathLike } from 'fs-extra';
 import { FSWatcher, watch } from 'chokidar';
+import { createHash } from 'crypto';
+import { diffArrays } from 'diff';
+import { PathLike } from 'fs-extra';
+import React, { ReactNode, createContext, useEffect } from 'react';
+import { flattenArray } from '../../containers/flatten';
 import useMap from '../../containers/hooks/useMap';
 import { WatchEventType } from '../../containers/hooks/useWatcher';
 import { extractStats, isEqualPaths, readFileAsync } from '../../containers/io';
+import { isDefined } from '../../containers/utils';
 import { useAppDispatch, useAppSelector } from '../hooks';
-import { isDirectoryMetafile, isFilebasedMetafile, isFileMetafile, metafileRemoved } from '../slices/metafiles';
 import cacheSelectors from '../selectors/cache';
-import { diffArrays } from 'diff';
-import { flattenArray } from '../../containers/flatten';
-import { cacheRemoved, cacheUpdated } from '../slices/cache';
-import { fetchCache, subscribe } from '../thunks/cache';
-import { fetchMetafile } from '../thunks/metafiles';
 import cardSelectors from '../selectors/cards';
 import metafileSelectors from '../selectors/metafiles';
-import { isDefined } from '../../containers/utils';
+import { cacheRemoved, cacheUpdated } from '../slices/cache';
 import { Card } from '../slices/cards';
+import { isDirectoryMetafile, isFileMetafile, isFilebasedMetafile, metafileRemoved } from '../slices/metafiles';
+import { fetchCache, subscribe } from '../thunks/cache';
+import { fetchMetafile } from '../thunks/metafiles';
 
 export const FSCache = createContext({});
 
@@ -58,12 +59,13 @@ export const FSCacheProvider = ({ children }: { children: ReactNode }) => {
                 const stats = await extractStats(filename);
                 const metafile = await dispatch(fetchMetafile({ path: filename })).unwrap();
                 const card = selectByFilepath(filename);
-                if (card && stats && !stats.isDirectory()) { // verify file exists on FS and is not a directory
+                const existing = await dispatch(fetchCache(filename)).unwrap();
+                if (!existing && card && stats && !stats.isDirectory()) { // verify file exists on FS and is not a directory
                     dispatch(subscribe({ path: filename, card: card.id }));
                     const newWatcher = watch(filename.toString(), { ignoreInitial: true });
                     newWatcher.on('all', eventHandler);
                     watcherActions.set(filename, newWatcher);
-                } else if (!stats) {
+                } else if (!existing && !stats) {
                     dispatch(metafileRemoved(metafile.id));
                 }
                 break;
@@ -72,7 +74,7 @@ export const FSCacheProvider = ({ children }: { children: ReactNode }) => {
                 const existing = await dispatch(fetchCache(filename.toString())).unwrap();
                 if (existing) dispatch(cacheUpdated({
                     ...existing,
-                    content: await readFileAsync(filename, { encoding: 'utf-8' })
+                    content: createHash('md5').update(await readFileAsync(filename, { encoding: 'utf-8' })).digest('hex')
                 }));
                 break;
             }
